requirements.txt
# Core Streamlit and web framework
streamlit>=1.28.0

# Deepgram SDK for speech recognition and analysis
deepgram-sdk>=4.8.1

# Template engine for narrative rendering
jinja2>=3.1.0

# HTTP requests for URL handling
requests>=2.31.0

# Audio processing and file handling
pydub>=0.25.1

# Environment variables
python-dotenv>=1.0.0

# Optional: For enhanced audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Optional: For better file type detection
python-magic>=0.4.27

# Optional: For progress bars and enhanced UI
tqdm>=4.65.0
